% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ll_convergence.R
\name{check_loglike_convergence}
\alias{check_loglike_convergence}
\title{Compute critical value for log-likelihood convergence test.}
\usage{
check_loglike_convergence(it, burnin, loglike)
}
\arguments{
\item{it}{Iteration number.}

\item{burnin}{Number of iterations before likelihood convergence criterion is
checked.}

\item{loglike}{Vector of log-likelihood values after each iteration.}
}
\value{
Critical value for log-likelihood convergence test.
}
\description{
Compute critical value for log-likelihood convergence test.
}
\examples{
check_loglike_convergence(2, 0, c(-100, -50))
}
