% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_to_core.R
\name{chunk_to_core}
\alias{chunk_to_core}
\title{Convert chunk_labs to core_labs, i.e. set all singleton chunklet labels to 0.}
\usage{
chunk_to_core(chunk_labs)
}
\arguments{
\item{chunk_labs}{Integer vector of chunklet labels.}
}
\value{
Integer vector of core_labels, same length as chunk_labs.
}
\description{
Convert chunk_labs to core_labs, i.e. set all singleton chunklet labels to 0.
}
