% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunklet_cores.R
\name{chunklet_cores}
\alias{chunklet_cores}
\title{Convert +/- Table Labels into Chunklets}
\usage{
chunklet_cores(data, table_labs, prob = 0.9)
}
\arguments{
\item{data}{Dataset in matrix or data.frame form.}

\item{table_labs}{Output from table_to_labs.}

\item{prob}{Probability cut-off for cores.}
}
\value{
A list of two label vectors, labs with all non-core points labelled 0,
chunks with all non-core points given their own chunklet label.
}
\description{
Convert +/- Table Labels into Chunklets
}
\examples{
iris_tab   <- rbind(se = c(-1, +1, -1, -1),
                    ve = c(00, -1, +1, +1),
                    vi = c(+1, 00, +1, +1))
iris_table_labs <- table_to_label(iris[, 1:4], type_marker = iris_tab)$labs
iris_chunk_labs <- chunklet_cores(iris[, 1:4], table_labs = iris_table_labs)
}
