% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_zones.R
\name{compute_splits}
\alias{compute_splits}
\title{Use flowDensity::deGate to split the markers.}
\usage{
compute_splits(data, type_marker)
}
\arguments{
\item{data}{Dataset in matrix or data.frame format.}

\item{type_marker}{Matrix, rows for populations, columns for variables.}
}
\value{
splits: a vector of bimodal threshold values for each marker
from flowDensity::deGate.
}
\description{
Use flowDensity::deGate to split the markers.
}
