% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constr_em2.R
\name{constr_em2}
\alias{constr_em2}
\title{Use the EM algorithm to fit a GMM that satisfies positive constraints.}
\usage{
constr_em2(
  data,
  clust_num,
  chunk_labs,
  maxit = 100,
  eps = 1e-10,
)
}
\arguments{
\item{data}{Dataset as a data.frame or matrix.}

\item{clust_num}{Number of components of GMM.}

\item{chunk_labs}{Vector giving the chunklet label of each observation.}

\item{maxit}{Maximum number of EM iterations.}

}
\value{
A list.
}
\description{
Use the EM algorithm to fit a GMM that satisfies positive constraints.
}
\examples{
constr_em2(iris[, 1:4], 3, 1:150)
}
