% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_kmeans.R
\name{constrained_kmeans}
\alias{constrained_kmeans}
\title{Runs fixed_kmeans given constraints_common}
\usage{
constrained_kmeans(data, constraints_common, clust_num, init_seed = 123)
}
\arguments{
\item{data}{.}

\item{constraints_common}{.}

\item{clust_num}{.}

\item{init_seed}{.}
}
\value{
label vector
}
\description{
Runs fixed_kmeans given constraints_common
}
