% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_zones.R
\name{construct_zones}
\alias{construct_zones}
\title{Construct positive / negative region labels.}
\usage{
construct_zones(data, type_marker)
}
\arguments{
\item{data}{Dataset in matrix or data.frame format.}

\item{type_marker}{Matrix, rows for populations, columns for variables.}
}
\value{
List containing labels, a matrix with rows for observations and
columns for populations, and splits, a vector of the variable bimodal
threshold values.
}
\description{
Convert a cell type vs marker table into a logical vector of region
membership for each observation.
}
