% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_kmeans.R
\name{fixed_kmeans}
\alias{fixed_kmeans}
\title{k-Means with some fixed centres provided.}
\usage{
fixed_kmeans(data, centres, clust_num, init_seed = 123, eps = 0.01)
}
\arguments{
\item{data}{.}

\item{centres}{.}

\item{clust_num}{.}

\item{init_seed}{.}

\item{eps}{.}
}
\value{
label vector
}
\description{
k-Means with some fixed centres provided.
}
