% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_model.R
\name{initialise_model}
\alias{initialise_model}
\title{Compute initial parameter values of a GMM.}
\usage{
initialise_model(data, clust_num, start = "vanilla")
}
\arguments{
\item{data}{Matrix or dataframe.}

\item{clust_num}{Number of clusters.}

\item{start}{Initialisation option.}
}
\value{
A list.
}
\description{
Compute initial parameter values of a GMM.
}
\examples{
initialise_model(iris[, 1:4], 4)
}
