% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_constraints.R
\name{label_constraints}
\alias{label_constraints}
\title{Produce constraint label vectors based on \code{zone_matrix}.}
\usage{
label_constraints(data, zone_matrix, zone_percent)
}
\arguments{
\item{data}{Dataset in \code{matrix} or \code{data.frame} format.}

\item{zone_matrix}{Logical matrix with a column per zone and a row per event.}

\item{zone_percent}{Percentage of events in zone to be included in each
constrained set, either one value for all constrained
sets or one value per constrained set.}
}
\value{
List:
\itemize{
\item common: vector of constrained set labels,
(unconstrained events have common label 0).
\item unique: vector of constrained set labels,
(unconstrained events have unique labels).
}
}
\description{
Produce constraint label vectors based on \code{zone_matrix}.
}
