% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_zones.R
\name{label_zones}
\alias{label_zones}
\title{Construct positive / negative region labels.}
\usage{
label_zones(data, type_marker)
}
\arguments{
\item{data}{Dataset in matrix or data.frame format.}

\item{type_marker}{Matrix, rows for populations, columns for variables.}
}
\value{
List containing labs, a matrix with rows for observations and columns for populations,
and splits, a vector of the variable bimodal threshold values.
}
\description{
Convert a cell type vs marker table into a logical vector of region
membership for each observation.
}
\examples{
iris_tab   <- rbind(se = c(-1, +1, -1, -1),
                    ve = c(00, -1, +1, +1),
                    vi = c(+1, 00, +1, +1))
out <- label_zones(iris[, 1:4], type_marker = iris_tab)
}
