% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mustlink.R
\name{mustlink}
\alias{mustlink}
\title{All-inclusive function for must-link GMM.}
\usage{
mustlink(
  data,
  clust_num,
  zone_matrix = NULL,
  zone_percent = 100,
  maxit = 10000,
  eps = 1e-10,
  init_seed = NULL,
  init_method = c("k-Means++", "k-Means", "Must-Link k-Means++", "Must-Link k-Means"),
  init_labels = NULL,
  print_freq = 10,
  burnin = 2,
  model = c("vm", "ns")
)
}
\arguments{
\item{data}{Dataset in matrix or data.frame format.}

\item{clust_num}{Number of clusters / components.}

\item{zone_matrix}{Logical matrix with a column per zone and a row per event.}

\item{zone_percent}{Percentage of events in zone to be included in each
chunklet, either one value for all chunklets or one value
per chunklet.}

\item{maxit}{Maximum number of EM iterations.}

\item{eps}{Convergence criterion for relative difference in log-likelihood.}

\item{init_seed}{Seed.}

\item{init_method}{Initialisation option.}

\item{init_labels}{Initial labels.}

\item{print_freq}{Controls how frequently the log-likelihood and time are
printed in the EM loop.}

\item{burnin}{Controls how many loops are completed before testing for
likelihood convergence.}

\item{model}{Model to be used. Either "vm" for Melnykov et al. or "ns" for
Shental et al.}
}
\value{
A list consisting of a vector of cluster labels,
a matrix of chunklet to cluster assignment probabilities,
a list of model parameters,
a vector of log-likelihood values,
and a vector of times.
}
\description{
Only requires flowFrame dataset, cell type - marker table, number of
clusters, and size of chunklets.
}
