% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mustlink_em.R
\name{mustlink_em}
\alias{mustlink_em}
\title{Must-Link / Positive Constraint EM GMM.}
\usage{
mustlink_em(
  data,
  clust_num,
  chunk_labs,
  maxit = 100,
  eps = 1e-10,
  start = "k-Means",
  init_seed = NULL
)
}
\arguments{
\item{data}{Dataset as a data.frame or matrix.}

\item{clust_num}{Number of components of GMM.}

\item{chunk_labs}{Vector giving the chunklet label of each observation.}

\item{maxit}{Maximum number of EM iterations.}

\item{eps}{Convergence criterion for relative difference in log-likelihood.}

\item{start}{Initialisation option.}

\item{init_seed}{Seed.}
}
\value{
A list consisting of a vector of cluster labels,
a matrix of chunklet to cluster assignment probabilities,
a list of model parameters,
a vector of log-likelihood values,
and a vector of times per iteration.
}
\description{
Use the EM algorithm to fit a Gaussian Mixture Model that satisfies
must-link / positive constraints.
}
\examples{
iris_frame <- flowCore::flowFrame(exprs = as.matrix(iris[, 1:4]))
iris_tab   <- rbind(se = c(-1, +1, -1, -1),
                    ve = c(00, -1, +1, +1),
                    vi = c(+1, 00, +1, +1))
iris_table_labs <- table_to_label(iris_frame, type_marker = iris_tab)$labs
iris_chunk_labs <- chunklet_cores(iris[, 1:4], table_labs = iris_table_labs)$chunks
mustlink_em(iris[, 1:4], clust_num = 3, chunk_labs = iris_chunk_labs)
}
