% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mustlink_em_vm.R
\name{mustlink_em_vm}
\alias{mustlink_em_vm}
\title{Must-Link / Positive Constraint EM GMM.}
\usage{
mustlink_em_vm(
  data,
  chunk_labels,
  params,
  clust_num,
  burnin = 10,
  maxit = 10000,
  eps = 1e-10,
  no_print = FALSE,
  print_freq = 1,
  model = "vm"
)
}
\arguments{
\item{data}{Dataset in matrix form.}

\item{chunk_labels}{Output from chunklet_cores.}

\item{params}{Model parameters, for example, output from initialise_model.}

\item{clust_num}{Number of clusters.}

\item{burnin}{Number of iterations before likelihood convergence criterion is
checked.}

\item{maxit}{Maximum number of EM iterations.}

\item{eps}{Likelihood convergence criterion threshold.}

\item{no_print}{If TRUE, no printing takes place.}

\item{print_freq}{Number of iterations between print statements during EM.}

\item{model}{Model to be used. Either "vm" for Melnykov et al. or "ns" for
Shental et al.}
}
\value{
List of chunklet posterior probability matrix, model parameters, and
vector of log-likelihood values for each iteration.
}
\description{
Must-Link / Positive Constraint EM GMM.
}
\examples{
iris_tab   <- rbind(se = c(-1, +1, -1, -1),
                    ve = c(00, -1, +1, +1),
                    vi = c(+1, 00, +1, +1))
iris_zone_labels <- label_zones(iris[, 1:4], type_marker = iris_tab)$labels
iris_chunk_labels <- label_chunklets(iris[, 1:4],
                                   zone_labels = iris_zone_labels,
                                   zone_percent = 90)$chunk
iris_init <- initialise_model(iris[, 1:4], clust_num = 3,
                              start = "k-Means", init_seed = 123)
mustlink_em_vm(as.matrix(iris[, 1:4]), clust_num = 3,
             chunk_labels = iris_chunk_labels, params = iris_init)
}
