% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mustlink_estep.R
\name{mustlink_estep}
\alias{mustlink_estep}
\title{E-Step for Must-link Constrained GMM.}
\usage{
mustlink_estep(
  data,
  block,
  params,
  event_num = nrow(data),
  var_num = ncol(params$mu),
  clust_num = nrow(params$mu),
  model = c("vm", "ns")
)
}
\arguments{
\item{data}{Dataset in matrix or data.frame format.}

\item{block}{Object from make_block.}

\item{params}{Model parameters, for example, output from initialise_model.}

\item{event_num}{Number of observations in the dataset.}

\item{var_num}{Number of varibles in the dataset.}

\item{clust_num}{Number of clusters / components.}

\item{model}{Model to be used. Either "vm" for Melnykov et al. or "ns" for
Shental et al.}
}
\value{
A list containing a log-likelihood value
and a posterior probability matrix.
}
\description{
Implement E-step of EM algorithm for GMM
with positive / must-link constraints.
}
