% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mustlink_estep_ns.R
\name{mustlink_estep_ns}
\alias{mustlink_estep_ns}
\title{E-Step for Must-link Constrained GMM.}
\usage{
mustlink_estep_ns(
  data,
  block,
  params,
  event_num = nrow(data),
  var_num = ncol(params$mu),
  clust_num = nrow(params$mu)
)
}
\arguments{
\item{data}{Dataset being clustered in matrix form.}

\item{block}{Object from make_block.}

\item{params}{List containing prop, mu, sigma.}

\item{event_num}{Number of observations in the dataset.}

\item{var_num}{Number of varibles in the dataset.}

\item{clust_num}{Number of clusters pre-specified.}
}
\value{
A list containing a log-likelihood value
and a posterior probability matrix.
}
\description{
Implement E-step of EM algorithm for GMM
with positive / must-link constraints.
}
\examples{
block_labels1 <- c(rep(1, 25), 2:26, rep(27, 25), 28:52, rep(53, 25), 54:78)
block1 <- list(labels = block_labels1,
               num = length(unique(block_labels1)),
               size = as.numeric(table(block_labels1)))
params1 <- initialise_model(iris[, 1:4], clust_num = 3)
mustlink_estep_ns(as.matrix(iris[, 1:4]), block = block1, params = params1,
                  event_num = 150, var_num = 4, clust_num = 3)
}
