% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mustlink_mstep.R
\name{mustlink_mstep}
\alias{mustlink_mstep}
\title{E-Step for Must-link Constrained GMM.}
\usage{
mustlink_mstep(
  data,
  postprob_event,
  postprob_block,
  block_num = nrow(postprob_block),
  clust_num = ncol(postprob_block),
  event_num = nrow(data),
  var_num = ncol(data),
  model = c("vm", "ns")
)
}
\arguments{
\item{data}{Dataset being clustered in matrix form.}

\item{postprob_event}{Expanded observation posterior probability matrix.}

\item{postprob_block}{block posterior probability matrix.}

\item{block_num}{Number of blocks.}

\item{clust_num}{Number of clusters pre-specified.}

\item{event_num}{Number of observations in the dataset.}

\item{var_num}{Number of variables in the dataset.}

\item{model}{Model to be used. Either "vm" for Melnykov et al. or "ns" for
Shental et al.}
}
\value{
List containing prop, mu, sigma.
}
\description{
Implement M-step of EM algorithm for GMM with positive / must-link
constraints.
}
