% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mustlink_mstep_ns.R
\name{mustlink_mstep_ns}
\alias{mustlink_mstep_ns}
\title{E-Step for Must-link Constrained GMM.}
\usage{
mustlink_mstep_ns(
  data,
  obs_pp,
  block_pp,
  block_num = nrow(block_pp),
  clust_num = ncol(block_pp),
  event_num = nrow(data),
  var_num = ncol(data)
)
}
\arguments{
\item{data}{Dataset being clustered in matrix form.}

\item{obs_pp}{Expanded observation posterior probability matrix.}

\item{block_pp}{Block posterior probability matrix.}

\item{block_num}{Number of blocks.}

\item{clust_num}{Number of clusters pre-specified.}

\item{event_num}{Number of observations in the dataset.}

\item{var_num}{Number of variables in the dataset.}
}
\value{
List containing prop, mu, sigma.
}
\description{
Implement M-step of EM algorithm for GMM with positive / must-link
constraints.
}
\examples{
block_labels1 <- c(rep(1, 25), 2:26, rep(27, 25), 28:52, rep(53, 25), 54:78)
block1 <- list(labels = block_labels1,
               num = length(unique(block_labels1)),
               size = as.numeric(table(block_labels1)))
params1 <- initialise_model(iris[, 1:4], clust_num = 3)
e_out1  <- mustlink_estep_ns(as.matrix(iris[, 1:4]),
                             block = block1, params = params1,
                             event_num = 150, var_num = 4, clust_num = 3)
obs_pp1 <- e_out1$obs_pp
block_pp1 <- e_out1$block_pp
mustlink_mstep_ns(as.matrix(iris[, 1:4]),
                  obs_pp = obs_pp1, block_pp = block_pp1)
}
