% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mustlinkMix-package.R
\docType{package}
\name{mustlinkMix-package}
\alias{mustlinkMix}
\alias{mustlinkMix-package}
\title{mustlinkMix: Must-Link Constrained Model-Based Clustering}
\description{
Gaussian Mixture Modelling clustering with positive / must-link constraints, as in Melnykov et al (2016). Code was originally based off the constrEM function in the MBCbook companion package to the book "Model-Based Clustering" by Bouveyron et al (2019).
}
\author{
\strong{Maintainer}: Ultán Doherty \email{dohertyu@tcd.ie} (\href{https://orcid.org/0000-0003-1791-395X}{ORCID})

}
\keyword{internal}
