% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_pdf.R
\name{norm_pdf}
\alias{norm_pdf}
\title{Compute the pdf of a multivariate Normal distribution.}
\usage{
norm_pdf(x, mu, sigma)
}
\arguments{
\item{x}{A numeric vector giving the value of a single observation or a
matrix where each row is an observation.}

\item{mu}{A numeric vector giving the mean of the distribution.}

\item{sigma}{A matrix giving the covariance matrix of the distribution.}
}
\value{
A numeric vector with one entry per observation provided.
}
\description{
Compute the pdf of a multivariate Normal distribution.
}
\examples{
value <- norm_pdf(rep(0, 4), rep(0, 4), diag(4))
}
