% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mustlink.R
\name{read_mustlink}
\alias{read_mustlink}
\title{Use read.table and readRDS to read files created by write.mustlink.}
\usage{
read_mustlink(
  file_prefix,
  clust_labs = TRUE,
  chunk_labs = TRUE,
  times = TRUE,
  em_ll = TRUE,
  em_chunk_pp = TRUE,
  em_params = TRUE
)
}
\arguments{
\item{file_prefix}{Desired file_prefix.}

\item{clust_labs}{Logical: should cluster labels be read from a file?}

\item{chunk_labs}{Logical: should chunklet labels be read from a file?}

\item{times}{Logical: should mustlink runtimes be read from a file?}

\item{em_ll}{Logical: should log-likelihood values be read from a file?}

\item{em_chunk_pp}{Logical: should posterior probability matrix be read from
a file?}

\item{em_params}{Logical: should model parameters be read from a file?}
}
\value{
Same format as mustlink function.
}
\description{
Use read.table and readRDS to read files created by write.mustlink.
}
