% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnorm_log_pdf.R
\name{unnorm_log_pdf}
\alias{unnorm_log_pdf}
\title{Compute unnormalised log pdf of a multivariate Normal distribution.}
\usage{
unnorm_log_pdf(x, mu, sigma_inv)
}
\arguments{
\item{x}{A matrix where each row is an observation.}

\item{mu}{Mean vector.}

\item{sigma_inv}{Inverse of the covariance matrix.}
}
\value{
A numeric vector with one entry per observation provided.
}
\description{
Compute unnormalised log pdf of a multivariate Normal distribution.
}
\examples{
unnorm_log_pdf(matrix(rep(2, 4), nrow = 2), mu = rep(1, 2), sigma_inv = diag(2))
}
