% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_to_core.R
\name{chunk_to_core}
\alias{chunk_to_core}
\title{Convert chunk_labels to core_labels, i.e. set all singleton chunklet labels to 0.}
\usage{
chunk_to_core(chunk_labels)
}
\arguments{
\item{chunk_labels}{Integer vector of chunklet labels.}
}
\value{
Integer vector of core_labels, same length as chunk_labels.
}
\description{
Convert chunk_labels to core_labels, i.e. set all singleton chunklet labels to 0.
}
