% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_f1.R
\name{compute_f1}
\alias{compute_f1}
\title{Compute f1 given two vectors of labels}
\usage{
compute_f1(
  clust_labels,
  true_labels,
  exclude_from_true = NULL,
  prec_rec = FALSE
)
}
\arguments{
\item{clust_labels}{Vector of cluster labels to evaluate.}

\item{true_labels}{Vector of true / reference labels to compare to.}

\item{exclude_from_true}{Vector of true labels to be excluded.
Ignored if NULL.}

\item{prec_rec}{Logical: Should vectors of precision and recall values,
with assignment based on f1, be returned?}
}
\value{
List of mean f1, mean precision, mean recall, and f1 matrix.
}
\description{
Compute f1 given two vectors of labels
}
