% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_f1.R
\name{compute_f1}
\alias{compute_f1}
\title{Compute f1 given two vectors of labels}
\usage{
compute_f1(clust_labs, true_labs)
}
\arguments{
\item{clust_labs}{Vector of cluster labels to evaluate.}

\item{true_labs}{Vector of true / reference labels to compare to.}
}
\value{
List of mean f1, mean precision, mean recall, and f1 matrix.
}
\description{
Compute f1 given two vectors of labels
}
