% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_parameters.R
\name{initial_parameters}
\alias{initial_parameters}
\title{Compute initial parameter values of a GMM from a vector of partition
labels.}
\usage{
initial_parameters(data, init_labels)
}
\arguments{
\item{data}{Matrix or dataframe.}

\item{init_labels}{Set of initial labels.}
}
\value{
A list consisting of a mixing proportions vector, a matrix of
component means, and an array containing component covariance
matrices.
}
\description{
Initialise a Gaussian Mixture Model.
}
