% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_partition.R
\name{initial_partition}
\alias{initial_partition}
\title{Construct an initial partition.}
\usage{
initial_partition(
  data,
  clust_num,
  linked_set_labels = NULL,
  init_seed = NULL,
  init_method = c("mlkmpp", "mlkm", "kmpp", "km", "old_mlkmpp", "old_mlkm")
)
}
\arguments{
\item{data}{Dataset in matrix or data.frame format.}

\item{clust_num}{Number of clusters / components.}

\item{linked_set_labels}{Label indicating which linked set an event
belongs to, 0 for unconstrained events.}

\item{init_seed}{Seed.}

\item{init_method}{Initialisation option.}
}
\value{
An integer vector.
}
\description{
Partition the data.
}
