% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_partition.R
\name{initial_partition}
\alias{initial_partition}
\title{Construct an initial partition.}
\usage{
initial_partition(
  data,
  clust_num,
  linked_set_labels = NULL,
  init_seed = NULL,
  init_method = c("Must-Link k-Means++", "k-Means++", "k-Means")
)
}
\arguments{
\item{data}{Matrix or dataframe.}

\item{clust_num}{Number of clusters.}

\item{linked_set_labels}{Label indicating which linked set an event
belongs to, 0 for unconstrained events.}

\item{init_seed}{Seed.}

\item{init_method}{Initialisation option. One of "Must-Link k-Means++",
"k-Means++", "k-Means", "Mclust", or "use_labels".}
}
\value{
An integer vector.
}
\description{
Partition the data.
}
\examples{
initialise_model(iris[, 1:4], 4)
}
