% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_model.R
\name{initialise_model}
\alias{initialise_model}
\title{Compute initial parameter values of a GMM.}
\usage{
initialise_model(clust_num, dimension)
}
\arguments{
\item{clust_num}{Number of clusters.}

\item{dimension}{Number of variables.}
}
\value{
A list.
}
\description{
Compute initial parameter values of a GMM.
}
\examples{
initialise_model(3, 4)
}
