% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_model.R
\name{initialise_model}
\alias{initialise_model}
\title{Compute initial parameter values of a GMM.}
\usage{
initialise_model(
  data,
  clust_num,
  labels = NULL,
  start = "k-Means",
  init_seed = NULL
)
}
\arguments{
\item{data}{Matrix or dataframe.}

\item{clust_num}{Number of clusters.}

\item{labels}{Set of initial labels.}

\item{start}{Initialisation option. One of "central", "k-Means", or "mclust".}

\item{init_seed}{Seed.}
}
\value{
A list consisting of a mixing proportions vector, a matrix of component means,
and an array containing component covariance matrices.
}
\description{
Initialise a Gaussian Mixture Model either with equal proportions, mean
vectors, and covariance matrices, or based on a k-Means clustering, or using
parameters computed by mclust.
}
\examples{
initialise_model(iris[, 1:4], 4)
}
