% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_model.R
\name{initialise_model}
\alias{initialise_model}
\title{Compute initial parameter values of a GMM.}
\usage{
initialise_model(
  data,
  clust_num,
  constraint_labels = NULL,
  init_labels = NULL,
  init_seed = NULL,
  init_method = c("Must-Link k-Means++", "k-Means++", "k-Means", "Mclust", "use_labels")
)
}
\arguments{
\item{data}{Matrix or dataframe.}

\item{clust_num}{Number of clusters.}

\item{constraint_labels}{Label indicating which constrained set an event
belongs to, 0 for unconstrained events.}

\item{init_labels}{Set of initial labels.}

\item{init_seed}{Seed.}

\item{init_method}{Initialisation option. One of "Must-Link k-Means++",
"k-Means++", "k-Means", "Mclust", or "use_labels".}
}
\value{
A list consisting of a mixing proportions vector, a matrix of
component means, and an array containing component covariance
matrices.
}
\description{
Initialise a Gaussian Mixture Model either with equal proportions, mean
vectors, and covariance matrices, or based on a k-Means clustering, or using
parameters computed by mclust.
}
\examples{
initialise_model(iris[, 1:4], 4)
}
