% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_constraints.R
\name{label_constraints}
\alias{label_constraints}
\title{Convert +/- Table Labels into Chunklets}
\usage{
label_constraints(data, zone_matrix, zone_percent)
}
\arguments{
\item{data}{Dataset in matrix or data.frame form.}

\item{zone_matrix}{Output from construct_zones.}

\item{zone_percent}{Percentage of events in zone to be included in each
chunklet, either one value for all chunklets or one value
per chunklet.}
}
\value{
A list of two label vectors, labels with all non-core points labelled
0, chunks with all non-core points given their own chunklet label.
}
\description{
Convert +/- Table Labels into Chunklets
}
