% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_chunk.R
\name{make_chunk}
\alias{make_chunk}
\title{Compute chunklet number, means, and sizes from labels.}
\usage{
make_chunk(data, chunk_labs)
}
\arguments{
\item{data}{Dataset to be clustered.}

\item{chunk_labs}{Vector of chunklet labels.}
}
\value{
A list.
}
\description{
Compute chunklet number, means, and sizes from labels.
}
\examples{
make_chunks(iris[, 1:4], c(rep(1, 25), 26:50, rep(51, 25), 76:100, rep(101, 25), 126:150))
}
