% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mustlink.R
\name{mustlink}
\alias{mustlink}
\title{All-inclusive function for must-link GMM.}
\usage{
mustlink(
  data,
  type_marker = NULL,
  clust_num,
  zone_percent = 100,
  maxit = 10000,
  eps = 1e-10,
  init_seed = NULL,
  init_method = c("Must-Link k-Means++", "k-Means++", "k-Means"),
  init_labels = NULL,
  print_freq = 10,
  burnin = 10,
  model = c("vm", "ns")
)
}
\arguments{
\item{data}{Dataset in matrix or data.frame format.}

\item{type_marker}{Matrix with entries +/-/0, rows for populations, columns
for variables.}

\item{clust_num}{Number of clusters / components.}

\item{zone_percent}{Percentage of events in zone to be included in each
chunklet, either one value for all chunklets or one value
per chunklet.}

\item{maxit}{Maximum number of EM iterations.}

\item{eps}{Convergence criterion for relative difference in log-likelihood.}

\item{init_seed}{Seed.}

\item{init_method}{Initialisation option.}

\item{init_labels}{Initial labels.}

\item{print_freq}{Controls how frequently the log-likelihood and time are
printed in the EM loop.}

\item{burnin}{Controls how many loops are completed before testing for
likelihood convergence.}

\item{model}{Model to be used. Either "vm" for Melnykov et al. or "ns" for
Shental et al.}
}
\value{
A list consisting of a vector of cluster labels,
a matrix of chunklet to cluster assignment probabilities,
a list of model parameters,
a vector of log-likelihood values,
and a vector of times.
}
\description{
Only requires flowFrame dataset, cell type - marker table, number of
clusters, and size of chunklets.
}
\examples{
iris_tab   <- rbind(se = c(-1, +1, -1, -1),
                    ve = c(00, -1, +1, +1),
                    vi = c(+1, 00, +1, +1))
mustlink(iris[, 1:4], type_marker = iris_tab,
         clust_num = 3, zone_percent = 90)
}
