% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mustlink.R
\name{mustlink}
\alias{mustlink}
\title{Fit a constrained Gaussian mixture model.}
\usage{
mustlink(
  data,
  clust_num,
  zone_matrix = NULL,
  zone_percent = 100,
  maxit = 10000,
  eps = 1e-10,
  init_seed = NULL,
  init_method = c("mlkmpp", "mlkm", "kmpp", "km", "old_mlkm", "old_mlkmpp"),
  init_labels = NULL,
  print_freq = 10,
  burnin = 2,
  drop_cluster = FALSE,
  prob_minimum = 0
)
}
\arguments{
\item{data}{Dataset in \code{matrix} or \code{data.frame} format.}

\item{clust_num}{Number of clusters / components.}

\item{zone_matrix}{Logical matrix with a column per zone and a row per event.}

\item{zone_percent}{Percentage of events in zone to be included in each
constrained set, either one value for all constrained
sets or one value per constrained set.}

\item{maxit}{Maximum number of EM iterations.}

\item{eps}{Convergence criterion for relative difference in log-likelihood.}

\item{init_seed}{Seed.}

\item{init_method}{Initialisation option.}

\item{init_labels}{Initial labels.}

\item{print_freq}{Controls how frequently the log-likelihood and time are
printed in the EM loop.}

\item{burnin}{Controls how many loops are completed before testing for
likelihood convergence.}

\item{drop_cluster}{Whether empty clusters should be dropped.}

\item{prob_minimum}{Remove any event whose component assignment probability
is less than this threshold and classify as an outlier.}
}
\value{
A list
\itemize{
\item clust_labels: vector of cluster labels
\item init_labels: vector of initial labels
\item constraints_common: vector of constrained set labels,
(unconstrained events have common label 0).
\item constraints_unique: vector of constrained set labels,
(unconstrained events have unique labels).
\item em: list, output from \code{mustlink_em}.
\item times: runtimes for setup, em, and labelling.
}
}
\description{
Given a logical \code{zone_matrix} indicating which events are in each region for
which a constrained set should be constructed, choose \code{zone_percent}\% of
those events to include in each constrained set.
}
