% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mustlink_em.R
\name{mustlink_em}
\alias{mustlink_em}
\title{Must-Link / Positive Constraint EM GMM.}
\usage{
mustlink_em(
  data,
  clust_num,
  chunk_labs,
  maxit = 100,
  eps = 1e-10,
  start = "k-Means"
)
}
\arguments{
\item{data}{Dataset as a data.frame or matrix.}

\item{clust_num}{Number of components of GMM.}

\item{chunk_labs}{Vector giving the chunklet label of each observation.}

\item{maxit}{Maximum number of EM iterations.}

\item{eps}{Convergence criterion for relative difference in log-likelihood.}

\item{start}{Initialisation option.}
}
\value{
A list consisting of a vector of cluster labels,
a matrix of chunklet to cluster assignment probabilities,
a list of model parameters,
a vector of log-likelihood values,
and a vector of times per iteration.
}
\description{
Use the EM algorithm to fit a Gaussian Mixture Model that satisfies
must-link / positive constraints.
}
\examples{
chunks_of_25 <- c(rep(1, 25), 2:26, rep(27, 25), 28:52, rep(53, 25), 54:78)
mustlink_em(iris[, 1:4], clust_num = 3, chunk_labs = chunks_of_25)
}
