% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mustlink_em.R
\name{mustlink_em}
\alias{mustlink_em}
\title{Must-Link / Positive Constraint EM GMM.}
\usage{
mustlink_em(
  data,
  block_labels,
  params,
  clust_num,
  zone_num,
  burnin = 2,
  maxit = 10000,
  eps = 1e-10,
  print_freq = 1,
  drop_cluster = FALSE
)
}
\arguments{
\item{data}{Dataset in matrix or data.frame format.}

\item{block_labels}{Each event in a particular linked set has the same
number and every non-linked event has its own number.}

\item{params}{Model parameters, for example, output from initialise_model.}

\item{clust_num}{Number of clusters / components.}

\item{zone_num}{Number of zones.}

\item{burnin}{Controls how many loops are completed before testing for
likelihood convergence.}

\item{maxit}{Maximum number of EM iterations.}

\item{eps}{Convergence criterion for relative difference in log-likelihood.}

\item{print_freq}{Controls how frequently the log-likelihood and time are
printed in the EM loop.}

\item{drop_cluster}{Should empty clusters be dropped.}
}
\value{
List of chunklet posterior probability matrix, model parameters, and
vector of log-likelihood values for each iteration.
}
\description{
Must-Link / Positive Constraint EM GMM.
}
