% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mustlink_em.R
\name{mustlink_em}
\alias{mustlink_em}
\title{Must-Link / Positive Constraint EM GMM.}
\usage{
mustlink_em(
  data,
  block_labels,
  params,
  clust_num,
  zone_num,
  burnin = 2,
  maxit = 10000,
  eps = 1e-10,
  print_freq = 1,
  model = c("vm", "ns"),
  drop_cluster = FALSE
)
}
\arguments{
\item{data}{Dataset in matrix form.}

\item{block_labels}{Each event in a particular linked set has the same
number and every non-linked event has its own number.}

\item{params}{Model parameters, for example, output from initialise_model.}

\item{clust_num}{Number of clusters.}

\item{zone_num}{Number of zones.}

\item{burnin}{Number of iterations before likelihood convergence criterion is
checked.}

\item{maxit}{Maximum number of EM iterations.}

\item{eps}{Likelihood convergence criterion threshold.}

\item{print_freq}{Number of iterations between print statements during EM.}

\item{model}{Model to be used. Either "vm" for Melnykov et al. or "ns" for
Shental et al.}
}
\value{
List of chunklet posterior probability matrix, model parameters, and
vector of log-likelihood values for each iteration.
}
\description{
Must-Link / Positive Constraint EM GMM.
}
