% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mustlink_mstep.R
\name{mustlink_mstep}
\alias{mustlink_mstep}
\title{E-Step for Must-link Constrained GMM.}
\usage{
mustlink_mstep(
  data,
  postprob_events,
  postprob_sets,
  sets_num = nrow(postprob_sets),
  clust_num = ncol(postprob_sets),
  event_num = nrow(data),
  var_num = ncol(data),
  drop_cluster = FALSE
)
}
\arguments{
\item{data}{Dataset in matrix or data.frame format.}

\item{postprob_events}{Expanded observation posterior probability matrix.}

\item{postprob_sets}{Posterior probability matrix for constrained sets.}

\item{sets_num}{Number of constrained sets including unconstrained events.}

\item{clust_num}{Number of clusters / components.}

\item{event_num}{Number of observations in the data set.}

\item{var_num}{Number of variables in the data set.}

\item{drop_cluster}{Should empty clusters be dropped.}
}
\value{
List:
\itemize{
\item prop
\item mu
\item sigma
}
}
\description{
Implement M-step of EM algorithm for GMM with positive / must-link
constraints.
}
