% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_pdf.R
\name{norm_pdf}
\alias{norm_pdf}
\title{Compute the pdf of a multivariate Normal distribution.}
\usage{
norm_pdf(x, mu, sigma_inv)
}
\arguments{
\item{x}{A numeric vector giving the value of a single observation or a
matrix where each row is an observation.}

\item{mu}{A numeric vector giving the mean of the distribution.}

\item{sigma_inv}{A matrix giving the inverse of the covariance matrix of the distribution.}
}
\value{
A numeric vector with one entry per observation provided.
}
\description{
Compute the pdf of a multivariate Normal distribution.
}
\examples{
norm_pdf(matrix(rep(2, 4), nrow = 2), mu = rep(1, 2), sigma_inv = diag(2))
}
