% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mustlink.R
\name{read_mustlink}
\alias{read_mustlink}
\title{Use read.table and readRDS to read files created by write.mustlink.}
\usage{
read_mustlink(
  file_prefix,
  clust_labels = TRUE,
  init_labels = TRUE,
  block_labels = TRUE,
  linked_set_labels = TRUE,
  times = TRUE,
  em_loglike = TRUE,
  em_postprob_block = TRUE,
  em_params = TRUE
)
}
\arguments{
\item{file_prefix}{Desired file_prefix.}

\item{clust_labels}{Logical: should cluster labels be read from a file?}

\item{init_labels}{Logical: should initial labels be read from a file?}

\item{block_labels}{Logical: should block labels be read from a file?}

\item{linked_set_labels}{Logical: should linked set labels be read from a
file?}

\item{times}{Logical: should mustlink runtimes be read from a file?}

\item{em_loglike}{Logical: should log-likelihood values be read from a file?}

\item{em_postprob_block}{Logical: should posterior probability matrix be read
from a file?}

\item{em_params}{Logical: should model parameters be read from a file?}
}
\value{
Same format as mustlink function.
}
\description{
Use read.table and readRDS to read files created by write.mustlink.
}
