% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequential_split.R
\name{sequential_split}
\alias{sequential_split}
\title{Convert +/- Table Labels into Chunklets}
\usage{
sequential_split(x, typemarker, plot = TRUE)
}
\arguments{
\item{x}{Dataset in matrix or data.frame form.}

\item{typemarker}{Cell-type marker table.}

\item{plot}{Logical value.}
}
\value{
splits, typemarker, subsetter
}
\description{
Convert +/- Table Labels into Chunklets
}
