% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_mustlink.R
\name{write_mustlink}
\alias{write_mustlink}
\title{Use write.table and saveRDS to save the output from a mustlink call.}
\usage{
write_mustlink(
  mustlink_out,
  file_prefix,
  clust_labels = TRUE,
  init_labels = TRUE,
  block_labels = TRUE,
  linked_set_labels = TRUE,
  times = TRUE,
  em_loglike = TRUE,
  em_postprob_block = TRUE,
  em_params = TRUE
)
}
\arguments{
\item{mustlink_out}{Output from mustlink function.}

\item{file_prefix}{Desired file_prefix.}

\item{clust_labels}{Logical: should cluster labels be written to a file?}

\item{init_labels}{Logical: should initial labels be written to a file?}

\item{block_labels}{Logical: should block labels be written to a file?}

\item{times}{Logical: should runtimes be written to a file?}

\item{em_loglike}{Logical: should log-likelihood values be written to a file?}

\item{em_postprob_block}{Logical: should posterior probability matrix be written to
a file?}

\item{em_params}{Logical: should model parameters be written to a file?}
}
\value{
Function does not have a return value.
}
\description{
Use write.table and saveRDS to save the output from a mustlink call.
}
